requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-community
langchain-huggingface
langchain-chroma
langchain-google-genai
chromadb
pymupdf
sentence-transformers
python-dotenv